requirements.txt
# Core dependencies
asyncio-nats-client>=0.11.0
httpx>=0.25.0
openai>=1.3.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# Database
asyncpg>=0.29.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
arxiv>=2.1.0

# Animation
manim>=0.18.0
Pillow>=10.0.0
numpy>=1.24.0

# Web framework (optional for API)
fastapi>=0.104.0
uvicorn>=0.24.0

# MCP Server
mcp>=1.0.0

# Utilities
rich>=13.7.0
tenacity>=8.2.0
